+++ MasterErp.Service/Shared/SharedService.cs
@@ -176,7 +176,7 @@ namespace MasterErp.Service.Shared
         {
             var results = Context.Jobs.Select(b => new SelectorDataModel
             {
-                Id = b.JobId,
+                Id = (int)b.JobId,
                 Name = b.NameAR,
             }).ToList();
             return results;
